+++ test/e2e/code/steps/files.js
@@ -91,28 +91,24 @@ When('I share the uploaded file in the room', async function() {
 });
 
 async function checkFileIsShared(chat) {
-    console.debug('CHECK IF FILE IS SHARED');
-    console.debug('WAITING FOR META');
     await this.waitFor(() => chat.metaLoaded);
-    console.debug('META LOADED');
-    console.debug('GET BY ID IN CHAT', this.uploadedFile.fileId, chat.id);
     const file = ice.fileStore.getByIdInChat(this.uploadedFile.fileId, chat.id);
-    console.debug('ENSURE LOADED');
     await file.ensureLoaded();
     expect(file.deleted).to.be.not.true;
     const messages = chat.messages;
-    await this.waitFor(() => messages.length === chat.isChannel ? 3 : 2);
+    await this.waitFor(() => messages.length === (chat.isChannel ? 4 : 2));
     const lastMessage = messages[messages.length - 1];
-    console.log('========================', lastMessage, '===================================');
     expect(lastMessage.files).deep.equal([this.uploadedFile.fileId]);
 }
 
-Then('Cucumbot can see the uploaded file in DM', function() {
+Then('Cucumbot can see the uploaded file in DM', async function() {
     const chat = ice.chatStore.directMessages[0];
+    await chat._messageHandler.getInitialPage(); // might not be active in multi-chat test after restart
     return checkFileIsShared.call(this, chat);
 });
-Then('Cucumbot can see the uploaded file in the room', function() {
+Then('Cucumbot can see the uploaded file in the room', async function() {
     const chat = ice.chatStore.channels[0];
+    await chat._messageHandler.getInitialPage();// might not be active in multi-chat test after restart
     return checkFileIsShared.call(this, chat);
 });